+++ csatar-plugins/csatar/lang/hu/lang.php
@@ -431,7 +431,8 @@
             ],
             'permissions' => [
                 'permissions' => 'Jogosultságok',
-                'allPermissionsForScout' => 'Minden jogosultság a Cserkész megbízáshoz, az összes model összes mezőjéhez.'
+                'allPermissionsForScout' => 'Minden jogosultság a Cserkész megbízáshoz, az összes model összes mezőjéhez.',
+                'readPermissionForGuests' => 'Olvasási jog vendégeknek.',
             ],
         ],
         'component' => [