+++ csatar-plugins/csatar/models/Patrol.php
@@ -25,7 +25,7 @@ class Patrol extends OrganizationBase
         'facebook_page' => 'url|regex:(facebook)|nullable',
         'logo' => 'image|nullable',
         'age_group' => 'required',
-         'team' => 'required',
+        'team' => 'required',
     ];
 
     /**
@@ -42,7 +42,7 @@ class Patrol extends OrganizationBase
         if ($this->troop_id && $this->troop->team->id != $this->team_id) {
             throw new \ValidationException(['troop' => Lang::get('csatar.csatar::lang.plugin.admin.patrol.troopNotInTheTeamError')]);
         }
-       
+
         // check that the required mandates are set for now
         $this->validateRequiredMandates($this->attributes);
     }
@@ -81,7 +81,12 @@ class Patrol extends OrganizationBase
      */
 
     public $belongsTo = [
-        'team' => '\Csatar\Csatar\Models\Team',
+        'team' => [
+            '\Csatar\Csatar\Models\Team',
+            'formBuilder' => [
+                'requiredBeforeRender' => true,
+            ],
+        ],
         'troop' => '\Csatar\Csatar\Models\Troop',
         'age_group' => '\Csatar\Csatar\Models\AgeGroup',
     ];
@@ -149,7 +154,7 @@ class Patrol extends OrganizationBase
         }
         return [];
     }
-    
+
     /**
      * Return all patrols, which belong to the given team
      */