+++ Changelog.md
@@ -2,11 +2,13 @@
 
 ## Unreleased
 
+- CS-273 - Added permissions handling to form builder
 - CS-257 - Added profile data tabs and profile data cards
 - CS-279 - Changed color of Grid based table heading
 - CS-278 - Changed the style of the explanatoy text
 - CS-277 - Modified the position of the flash messages, depending on the screen size
 - CS-47 - Created basics of the rights system
+- CS-47  - Created basics of the rights system
 - CS-268 - Added new pictograms, and changed definition of existing ones
 - CS-199 - Extended social login plugin to handle Scout and User model connections
 - CS-193 - Applied basic permissions